Fix flat demand readings
Flat stretches in Wh was caused by filling for missing days. Because
some buildings have prolonged periods without data (2-3 months), before
data recommences the existing filling code would create large stretches
of the same value for all variables. This sometimes resulted in
`model.matrix` throwing errors. This filling approach was only supposed
to be for small gaps in the data. Now filling is only done for predictor
variables. If the response is missing from a row it is now filtered.